RVEnet/utils/fix_seed.py
import torch
import numpy as np
import random

def fix_seed(seed: int):

    torch.manual_seed(seed)
    torch.cuda.manual_seed_all(seed)
    torch.backends.cudnn.deterministic = True
    torch.backends.cudnn.benchmark = False

    random.seed(seed)
    np.random.seed(seed)