requirements.txt
matplotlib==3.4.3
numpy==1.21.4
opencv_contrib_python==4.5.5.64
pandas==1.3.4
Pillow==10.1.0
pydicom==2.0.0
tqdm==4.60.0
utils==1.0.1
